requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.0.1
passlib[bcrypt]
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
playwright==1.49.1
pluggy==1.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.0.0
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.6.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.41.3
text-unidecode==1.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
